+++ packages/core/src/lib/model.spec.ts
@@ -1862,7 +1862,7 @@ describe("Test Model", () => {
     it("should throw error when trying to updateMultiple on models with allowMultipleOperations = false", async () => {
       const TestModel = mockModel().extend({ adapterClass: mockAdapter() });
 
-      TestModel.configuration.blockMultipleOperations = true;
+      TestModel.configuration.noBulk = true;
 
       await expect(TestModel.update({}, {})).rejects.toThrow("Cannot run updateMultiple operation");
     });
@@ -1870,7 +1870,7 @@ describe("Test Model", () => {
     it("should be able to updateMultiple on models with allowMultipleOperations = false and query as string (=updateOne)", async () => {
       const TestModel = mockModel().extend({ adapterClass: mockAdapter() });
 
-      TestModel.configuration.blockMultipleOperations = true;
+      TestModel.configuration.noBulk = true;
 
       await expect(TestModel.update("", {})).resolves.toBeDefined();
     });
@@ -1878,7 +1878,7 @@ describe("Test Model", () => {
     it("should throw error when trying to deleteMultiple on models with allowMultipleOperations = false", async () => {
       const TestModel = mockModel().extend({ adapterClass: mockAdapter() });
 
-      TestModel.configuration.blockMultipleOperations = true;
+      TestModel.configuration.noBulk = true;
 
       await expect(TestModel.delete({})).rejects.toThrow("Cannot run deleteMultiple operation");
     });
@@ -1886,7 +1886,7 @@ describe("Test Model", () => {
     it("should be able to deleteMultiple on models with allowMultipleOperations = false and query as string (=deleteOne)", async () => {
       const TestModel = mockModel().extend({ adapterClass: mockAdapter() });
 
-      TestModel.configuration.blockMultipleOperations = true;
+      TestModel.configuration.noBulk = true;
 
       await expect(TestModel.delete("")).resolves.toBeDefined();
     });